Adds 'get_data' method to traverse scope chain
This adds a method to look up keys. We could just do this with a nice
hash, but we need to be able to add or remove items from the scope
without stomping on scopes outside it.

The spec did not explicitly state that inside scopes should be able
to use keys for outside scopes, but I consider it rude not to support
those semantics.