Support EACH
This adds 'each' support to the parser. We now parse a tree rather
than just a flat list. Anything inside an EACH/ENDEACH block will
end up as a child node of that EACH block. This made some changes
to the parser necessary.

It would probably be somewhat easy to make this whole system work
in just one pass.